backend/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.8
requests==2.22.0
google-auth==1.6.3
